requirements.txt
aiodns==3.2.0
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
alembic==1.14.0
Babel==2.14.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
frozenlist==1.4.1
greenlet==3.0.3
hiredis==2.3.2
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pycares==4.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
redis==5.0.4
SQLAlchemy==2.0.29
typing_extensions==4.11.0
uvloop==0.19.0
yarl==1.9.4
aiodns==3.2.0
aiofiles==23.2.1
aiogram==3.5.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.6.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.14.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
frozenlist==1.4.1
greenlet==3.0.3
hiredis==2.3.2
idna==3.7
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.0.5
pycares==4.4.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
redis==5.0.4
SQLAlchemy==2.0.29
typing_extensions==4.11.0
uvloop==0.19.0
yarl==1.9.4
